moved enums to a new directory
