+++ conveyor/src/main/java/com/neoflex/conveyor/dto/enums/EmploymentStatusType.java
@@ -1,4 +1,4 @@
-package com.neoflex.conveyor.dto.enumType;
+package com.neoflex.conveyor.dto.enums;
 
 public enum EmploymentStatusType {
     EMPLOYED,
