+++ conveyor/src/main/java/com/neoflex/conveyor/dto/enums/GenderType.java
@@ -1,4 +1,4 @@
-package com.neoflex.conveyor.dto.enumType;
+package com.neoflex.conveyor.dto.enums;
 
 public enum GenderType {
     MALE,
